Replace operatingsystem with osfamily in params.pp
With the latest version of Puppet, I am running
into an issue where the ubuntu operatingsystem does
not match the previous existing regular expression.

This commit changes the more complicated regular expressions
with the usage of osfamily. This resolves an issue that I
am seeing with Puppet 3.2.3 and also makes this module
more consistent with the other openstack modules.